reorganize directory: move some wish routes to user
